+++ observatory-dags/observatory/dags/telescopes/google_analytics.py
@@ -32,8 +32,8 @@ from oauth2client.service_account import ServiceAccountCredentials
 from observatory.api.client.model.organisation import Organisation
 from observatory.platform.telescopes.snapshot_telescope import SnapshotRelease, SnapshotTelescope
 from observatory.platform.utils.airflow_utils import AirflowConns, AirflowVars
-from observatory.platform.utils.telescope_utils import list_to_jsonl_gz, make_dag_id
-from observatory.platform.utils.template_utils import add_partition_date, blob_name, bq_load_partition, \
+from observatory.platform.utils.telescope_utils import add_partition_date, list_to_jsonl_gz, make_dag_id
+from observatory.platform.utils.template_utils import blob_name, bq_load_partition, \
     table_ids_from_path
